LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animation-base-response-004-expected.txt


PASS Base is responsive to font-affecting appearing via setKeyframes
